server_side/book_inventory/user/urls.py
from django.urls import path
from .views import *
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView


urlpatterns = [
    path('', UserDetailView.as_view(), name='user_detail'),
    path('register/', UserRegistrationView.as_view(), name='user_registration'),
    # jwt authentication urls
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
